set IPV6_V6ONLY on the coap socket
some operating systems (bsd, some debian based systems[1]), sockets have
to explicitly request ipv6-wrapped ipv4 packages (eg. from
::ffff:192.168.0.1). this patch makes smcp request those if the
setsockopt flag to do so is defined, as this is the only way for the
code to deal with ipv4 packages.

see [1] for a discussion on the default values of this option.

[1] http://bugs.debian.org/cgi-bin/bugreport.cgi?bug=560238