Major rework and start of 0.07.00 development.
Originally I had intended for this commit to include real
session support. But after struggling to come up with a
sane interface, I realized that the best way forward is to
enable session management at the platform level rather than
add session management directly to SMCP. This keeps implementations
small for applications which don't need full session capability.

The big win in this release is that the platform API is now
much more clearly defined. Every function that you see that
starts with `smcp_plat_` is a platform-defined function.
This makes it easier to figure out what functions you need
to implement for supporting various platforms.

A big focus of the 0.07.00 release is trying to achieve the
semblance of API stability. Still a long way away from that
goal, but this commit makes significant progress toward that
end.