smcpd: Now useful again.
The primary motivation behind this commit is to get `smcpd` back
up and running again. Yes, it is still very crusty code, but
it is an ongoing priority to clean it up. Small steps.

This change also renames the confusingly named `smcp-variable-node`
object to the `smtp-variable-handler` object.

In the cases where we are dealing with `select()`, we now use
the `fd_count` instead of `max_fd`. The difference is that
`fd_count` represents the one plus the value of the last fd, compared
to `max_fd` that directly represents the last fd. This is to bring
it in-line with the actual behavior of `select()`.