Experimental DTLS Support via OpenSSL
This commit enables initial experimental support for DTLS for
`plat-bsd`, via OpenSSL. Since it is experimental and still
a bit tempermental, you must explicitly enable this support
using the `--enable-dtls` configure script flag.

A great deal of additional work is going to be needed before
we can remove the "experimental" qualifier from the name of
this feature, but progress is progress.

This commit is a part of issue #35.