Transaction max resends control
This commit adds some fields to a transaction that enables control
the max number of messages that will be sent after the transaction
receives a response.

This is useful to allow multicast transactions (that keep sending
requests after receives a response) to stop (or limit) resending,
but keep waiting for other responses.